# nexus/loomctl/openapi.yaml
openapi: 3.0.3
info:
  title: LoomCtl API
  description: |
    **LoomOS Control Plane API**
    
    Advanced enterprise-grade control plane for LoomOS providing:
    - Distributed AI training orchestration
    - Real-time job management and monitoring
    - Multi-tenant resource allocation
    - WebSocket streaming for live updates
    - Marketplace integration for AI services
    - TEE attestation and security
    
    ## Authentication
    All endpoints require Bearer token authentication.
    
    ## Rate Limiting
    - 1000 requests per minute for standard endpoints
    - 100 requests per minute for compute-intensive operations
    
    ## WebSocket Connections
    Real-time updates available via WebSocket at `/ws/{client_id}`
    
  version: 1.0.0
  contact:
    name: LoomOS Support
    email: support@loomos.ai
    url: https://loomos.ai/docs
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.loomos.ai/v1
    description: Production server
  - url: https://staging-api.loomos.ai/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  schemas:
    JobSpec:
      type: object
      required:
        - job_type
        - name
        - image
      properties:
        job_id:
          type: string
          description: Unique job identifier (auto-generated if not provided)
          example: "job_abc12345"
        job_type:
          type: string
          enum: [training, inference, evaluation, data_processing, model_adaptation, verification]
          description: Type of job to execute
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable job name
          example: "BERT Fine-tuning on Custom Dataset"
        description:
          type: string
          description: Optional job description
          example: "Fine-tune BERT model on customer support tickets"
        image:
          type: string
          description: Container image to run
          example: "loomos/pytorch:latest"
        command:
          type: array
          items:
            type: string
          description: Command to execute in container
          example: ["python", "train.py", "--epochs", "10"]
        environment:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
          example:
            CUDA_VISIBLE_DEVICES: "0,1"
            LEARNING_RATE: "0.001"
        resources:
          type: array
          items:
            $ref: '#/components/schemas/JobResource'
          description: Resource requirements
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 1
          description: Job priority (1=lowest, 10=highest)
        timeout_seconds:
          type: integer
          minimum: 1
          description: Maximum job execution time
          example: 3600
        retry_limit:
          type: integer
          minimum: 0
          maximum: 10
          default: 3
          description: Maximum retry attempts
        world_size:
          type: integer
          minimum: 1
          default: 1
          description: Number of workers for distributed training
        compression_ratio:
          type: number
          minimum: 0.001
          maximum: 1.0
          default: 0.01
          description: Gradient compression ratio for distributed training
        tags:
          type: object
          additionalProperties:
            type: string
          description: User-defined tags
          example:
            project: "nlp_research"
            team: "ai_platform"
...
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          description: Unique job identifier
          example: "job_abc12345"
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled, paused]
          description: Current job status
        progress:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Job completion progress (0.0-1.0)
          example: 0.75
        message:
          type: string
          description: Status message or latest update
          example: "Training epoch 8/10 completed"
        created_at:
          type: string
          format: date-time
          description: Job creation timestamp
        started_at:
          type: string
          format: date-time
          description: Job start timestamp
        completed_at:
          type: string
          format: date-time
          description: Job completion timestamp
        result:
          type: object
          description: Job execution results
          example:
            model_accuracy: 0.95
            training_loss: 0.12
            model_path: "/models/bert_finetuned.pth"
        error:
          type: string
          description: Error message if job failed
          example: "CUDA out of memory"
        resources_used:
          type: array
          items:
            $ref: '#/components/schemas/JobResource'
          description: Actual resources consumed
        assigned_workers:
          type: array
          items:
            type: string
          description: Workers assigned to this job
          example: ["worker_001", "worker_002"]
    WorkerInfo:
      type: object
      properties:
        worker_id:
          type: string
          description: Unique worker identifier
          example: "worker_gpu_001"
        status:
          type: string
          enum: [idle, busy, offline, maintenance]
          description: Current worker status
        capabilities:
          type: object
          description: Worker capabilities and resources
          example:
            gpu_count: 2
            gpu_memory_gb: 32
            cpu_cores: 16
            memory_gb: 64
        current_load:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Current resource utilization (0.0-1.0)
          example: 0.3
        last_heartbeat:
          type: string
          format: date-time
          description: Last heartbeat timestamp
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Job not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: "trace_xyz789"
paths:
...
  /jobs:
    post:
      summary: Submit Job
      description: Submit a new job for execution
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSpec'
            examples:
              training_job:
                summary: ML Training Job
                value:
                  job_type: "training"
                  name: "BERT Fine-tuning"
                  description: "Fine-tune BERT on custom dataset"
                  image: "loomos/pytorch:latest"
                  command: ["python", "train.py"]
                  resources:
                    - resource_type: "gpu"
                      amount: 2
                      unit: "cards"
                    - resource_type: "memory"
                      amount: 32
                      unit: "GB"
                  world_size: 4
                  compression_ratio: 0.01
              inference_job:
                summary: Inference Job
                value:
                  job_type: "inference"
                  name: "Batch Inference"
                  image: "loomos/inference:latest"
                  command: ["python", "infer.py", "--batch-size", "64"]
                  resources:
                    - resource_type: "gpu"
                      amount: 1
                      unit: "cards"
      responses:
        '200':
          description: Job submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    example: "job_abc12345"
                  status:
                    type: string
                    example: "submitted"
        '400':
          description: Invalid job specification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
    
    get:
      summary: List Jobs
      description: List jobs for the authenticated user
      tags: [Jobs]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled, paused]
          description: Filter by job status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of jobs to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of jobs to skip
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobStatus'
        '401':
          description: Unauthorized
  /jobs/{job_id}:
    get:
      summary: Get Job Status
      description: Get detailed status of a specific job
      tags: [Jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Unique job identifier
          example: "job_abc12345"
      responses:
        '200':
          description: Job status details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Cancel Job
      description: Cancel a running or pending job
      tags: [Jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Unique job identifier
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Job cancelled successfully"
        '400':
          description: Job cannot be cancelled
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Job not found
  /jobs/{job_id}/logs:
    get:
      summary: Stream Job Logs
      description: Stream real-time logs from a running job
      tags: [Jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Unique job identifier
      responses:
        '200':
          description: Server-sent events stream of log entries
          content:
            text/plain:
              schema:
                type: string
                example: |
                  data: {"timestamp": "2025-09-23T10:30:00Z", "level": "INFO", "message": "Training started"}
                  
                  data: {"timestamp": "2025-09-23T10:30:30Z", "level": "INFO", "message": "Epoch 1/10 completed"}
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Job not found
  /workers:
    get:
      summary: List Workers
      description: List all active workers in the cluster
      tags: [Workers]
      responses:
        '200':
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerInfo'
        '401':
          description: Unauthorized
...
  /cluster/stats:
    get:
      summary: Cluster Statistics
      description: Get comprehensive cluster statistics
      tags: [Cluster]
      responses:
        '200':
          description: Cluster statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterStats'
        '401':
          description: Unauthorized
...